+++ vhost_user_block/src/lib.rs
@@ -153,8 +153,9 @@ impl VhostUserBlkThread {
             }
 
             if self.event_idx {
-                if let Some(used_idx) = vring.mut_queue().add_used(mem, head.index, len) {
-                    if vring.needs_notification(&mem, Wrapping(used_idx)) {
+                let queue = vring.mut_queue();
+                if let Some(used_idx) = queue.add_used(mem, head.index, len) {
+                    if queue.needs_notification(&mem, Wrapping(used_idx)) {
                         debug!("signalling queue");
                         vring.signal_used_queue().unwrap();
                     } else {
